+++ fetchers/test_all/run.py
@@ -43,661 +43,6 @@ driver = webdriver.Firefox(
     executable_path="/usr/local/share/geckodriver", options=firefox_options)
 
 
-def manta_data_scraper(url, source_phone):
-    driver_2 = webdriver.Chrome(chrome_options=chrome_options)
-    import json
-    print(space)
-    new_data_list = []
-    validated = False
-    url = "https://" + url
-    print(url)
-    website = ""
-    try:
-        driver_2.get(url)
-    except:
-        raw_input("Driver isn't working properly press Enter to continue.")
-        driver_2.get(url)
-    html_page = driver_2.page_source
-    page_soup = soup(html_page, 'html.parser')
-    if "Access denied" in page_soup.text:
-        print("Blocked")
-    if "One more step" in page_soup.text:
-        raw_input("Captcha")
-    try:
-        claimed = page_soup.find("div", {
-                                 "class": "inline-block mt-1 rounded-r bg-gray-light h-7 px-2 py-1 text-xs text-gray-dark"}).text.strip().lower()
-        claimed = string_cleaner(claimed)
-        print("Claimed: " + claimed)
-    except:
-        claimed = ""
-    one = page_soup.find("script", type="application/ld+json")
-    two = str(one).split('<script type="application/ld+json">')[1]
-    three = two.split('</script>')[0]
-    json_snippet = json.loads(three)
-    try:
-        company_name = json_snippet["name"]
-        company_name = string_cleaner(company_name)
-        print("Company Name: " + company_name)
-    except:
-        company_name = ""
-    try:
-        phone_number = json_snippet["telephone"]
-        phone_number = remove_phone_format(phone_number)
-        if phone_number == source_phone:
-            validated = True
-        print("Phone Number: " + phone_number)
-    except:
-        phone_number = ""
-    try:
-        email_address = json_snippet["email"]
-        email_address = string_cleaner(email_address)
-        if "null" in email_address:
-            email_address = ""
-        else:
-            print("Email: " + email_address)
-    except:
-        email_address = ""
-    try:
-        street_address = str(json_snippet["address"]["streetAddress"][0])
-        street_address = string_cleaner(street_address)
-        print("Street: " + street_address)
-    except:
-        street_address = ""
-    try:
-        city = str(city["address"]["addressLocality"][0])
-        city = string_cleaner(city)
-        print("City: " + city)
-    except:
-        city = ""
-    try:
-        state = str(json_snippet["address"]["addressRegion"])
-        state = string_cleaner(state)
-        print("State: " + state)
-    except:
-        state = ""
-    try:
-        zip_code = str(json_snippet["address"]["postalCode"])
-        zip_code = string_cleaner(zip_code)
-    except:
-        zip_code = ""
-    else:
-        print("Zip Code: " + zip_code)
-    try:
-        country = str(country["address"]["addressCountry"][0])
-        country = string_cleaner(country)
-        print("Country: " + country)
-    except:
-        country = ""
-    try:
-        contact_name = str(json_snippet["employee"]["givenName"])
-        contact_name = string_cleaner(contact_name)
-        print("First Name: " + contact_name)
-    except:
-        contact_name = ""
-    try:
-        contact_last_name = str(json_snippet["employee"]["familyName"])
-        contact_last_name = string_cleaner(contact_last_name)
-        print("Last Name: " + contact_last_name)
-    except:
-        contact_last_name = ""
-    try:
-        contact_title = json_snippet["employee"]["jobTitle"]
-        contact_title = string_cleaner(contact_title)
-        print("Title: " + contact_title)
-    except:
-        contact_title = ""
-    try:
-        num_employees = str(json_snippet["numberOfEmployees"])
-        num_employees = string_cleaner(num_employees)
-        print("Employees: " + num_employees)
-    except:
-        num_employees = ""
-    try:
-        years = int(json_snippet["foundingDate"])
-        years = datetime.now().year-years
-        years = string_cleaner(years)
-        print("Years in business: " + str(years))
-    except:
-        years = ""
-    try:
-        stars = str(json_snippet["aggregateRating"]["ratingValue"])
-        stars = string_cleaner(stars)
-        print("Rating: " + stars)
-    except:
-        stars = ""
-    try:
-        reviews = str(json_snippet["aggregateRating"]["reviewCount"])
-        reviews = string_cleaner(reviews)
-        print("Reviews: " + reviews)
-    except:
-        try:
-            reviews = page_soup.find("div", {"class": "mbm"})
-            reviews = reviews.find("a").text.split(" ")[0]
-            reviews = string_cleaner(reviews)
-            print("Reviews: " + reviews)
-        except:
-            reviews = ""
-    try:
-        website_l = json_snippet["sameAs"]
-        try:
-            if isinstance(website_l, list):
-                for w in website_l:
-                    website = website + \
-                        str(w).replace("http://", "").replace("www.",
-                                                              "").replace("https://", "") + " "
-                    website = string_cleaner(website)
-            else:
-                website = str(website_l).replace(
-                    "http://", "").replace("www.", "").replace("https://", "")
-                website = string_cleaner(website)
-        except:
-            pass
-    except:
-        website = ""
-    else:
-        if "null" in website:
-            website = ""
-        else:
-            print("Website : " + website)
-    try:
-        category_container = page_soup.find(
-            "div", {"class": "text-gray-400 text-sm py-3 hidden lg:block"})
-        category_container = category_container.text.strip()
-        category = category_container.split("\n")[-1]
-        string_cleaner(category)
-        print("Category: " + category)
-    except:
-        category = ""
-    print("Validated: " + str(validated))
-    driver_2.close()
-    new_data_list.append(claimed)
-    new_data_list.append(company_name)
-    new_data_list.append(phone_number)
-    new_data_list.append(email_address)
-    new_data_list.append(street_address)
-    new_data_list.append(city)
-    new_data_list.append(state)
-    new_data_list.append(zip_code)
-    new_data_list.append(contact_name)
-    new_data_list.append(contact_last_name)
-    new_data_list.append(contact_title)
-    new_data_list.append(num_employees)
-    new_data_list.append(years)
-    new_data_list.append(stars)
-    new_data_list.append(reviews)
-    new_data_list.append(website)
-    new_data_list.append(category)
-    new_data_list.append(validated)
-    return new_data_list
-
-
-def mapquest_data_scraper(url, source_phone):
-    print(space)
-    print(url)
-    new_data_list = []
-    validated = False
-    try:
-        driver.get("https://" + url)
-    except:
-        raw_input("Driver isn't working properly. Press Enter to continue")
-        driver.get("https://" + url)
-    html_page = driver.page_source
-    page_soup = soup(html_page, 'html.parser')
-    try:
-        error_container = page_soup.find("body", {"class": "error-page"})
-        if "WHOOPS!" in error_container.text:
-            print("error page")
-    except:
-        pass
-    try:
-        company_name = page_soup.find(
-            "div", {"class": "header-wrapper"}).text.strip()
-    except:
-        company_name = ""
-    else:
-        try:
-            company_name = string_cleaner(company_name)
-            print("Company Name: " + company_name)
-        except:
-            pass
-    try:
-        email = page_soup.find("email", {"class": "ng-scope"})
-        email = email.find("a", {"class": "ng-scope"}
-                           )["href"].replace("mailto:", "")
-        email = string_cleaner(email)
-        print("Email: " + email)
-    except:
-        email = ""
-    try:
-        reviews = page_soup.find(
-            "span", {"class": "numerals ng-binding"}).text.strip().replace(" Reviews", "")
-        reviews = string_cleaner(reviews)
-        print("Reviews: " + reviews)
-    except:
-        reviews = ""
-    try:
-        stars = page_soup.find("meta", {"itemprop": "ratingValue"})
-        stars = stars.text.strip()
-        stars = string_cleaner(stars)
-        print("Rating: " + stars)
-    except:
-        stars = ""
-    try:
-        claimed_container = page_soup.find("span", {"id": "verified-business"})
-        claimed = claimed_container.text.strip()
-        claimed = string_cleaner(claimed)
-        print("Claimed: " + claimed)
-    except:
-        claimed = ""
-    try:
-        cc_payment = page_soup.find("li", {"itemprop": "paymentAccepted"})
-        cc_payment = cc_payment.text.strip()
-        cc_payment = string_cleaner(cc_payment)
-        print("Accepted Payment: " + cc_payment)
-    except:
-        cc_payment = ""
-    try:
-        phone_number = page_soup.find(
-            "p", {"ng-if": "ctrl.getPhone()"}).text.strip()
-        phone_number = remove_phone_format(phone_number)
-        print("Phone Number: " + phone_number)
-        if phone_number == source_phone:
-            validated = True
-    except:
-        phone_number = ""
-    print("Validated: " + str(validated))
-    new_data_list.append(company_name)
-    new_data_list.append(email)
-    new_data_list.append(reviews)
-    new_data_list.append(stars)
-    new_data_list.append(claimed)
-    new_data_list.append(cc_payment)
-    new_data_list.append(phone_number)
-    new_data_list.append(validated)
-    return new_data_list
-
-
-def chamberofcommerce_data_scraper(url, source_phone):
-    import urllib2
-    import time
-    hdr = {'User-Agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.11 (KHTML, like Gecko) Chrome/23.0.1271.64 Safari/537.11',
-           'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
-           'Accept-Charset': 'ISO-8859-1,utf-8;q=0.7,*;q=0.3',
-           'Accept-Encoding': 'none',
-           'Accept-Language': 'en-US,en;q=0.8',
-           'Connection': 'keep-alive'}
-    validated = False
-    new_data_list = []
-    print(space)
-    url = "https://" + url
-    print(url)
-    try:
-        req = urllib2.Request(url, headers=hdr)
-        try:
-            page = urllib2.urlopen(req)
-        except:
-            print("Check your url string and try again")
-        else:
-            time.sleep(1)
-            if page.getcode() != 200:
-                print("unsuccessful request")
-    except urllib2.HTTPError as e:
-        print("error retreiving html")
-    else:
-        page_soup = soup(page, 'html.parser')
-        try:
-            company_name = page_soup.find(
-                "div", {"class": "profile_business_name"}).text.strip()
-            company_name = string_cleaner(company_name)
-            print("Company Name: " + company_name)
-        except:
-            company_name = ""
-        try:
-            phone_number = page_soup.find(
-                "span", {"class": "d-none d-sm-block phone-align"}).text.strip()
-            phone_number = phone_number.replace("+1-", "")
-            phone_number = remove_phone_format(phone_number)
-            if phone_number == source_phone:
-                validated = True
-            print("Phone Number: " + phone_number)
-        except:
-            phone_number = ""
-        try:
-            reviews = page_soup.find(
-                "span", {"class": "review_conts_profile"}).text.strip()
-            reviews = reviews.replace("(", "")
-            reviews = reviews.replace(")", "")
-            reviews = string_cleaner(reviews)
-            print("Reviews: " + reviews)
-        except:
-            reviews = ""
-        try:
-            rating = page_soup.find("div", {"class": "review_rating"})
-            rating = str(rating)
-            rating = rating.split("data-rateit-value")[1]
-            rating = rating.split("data-rateit.ispreset")[0]
-            rating = rating.replace('"', '')
-            rating = rating.replace('=', '')
-            rating = string_cleaner(rating)
-            print("Rating: " + rating)
-        except:
-            rating = ""
-        try:
-            contact_container = page_soup.find(
-                "div", {"class": "about_p_text"})
-            contact_container = str(contact_container)
-            about_sections = contact_container.split("<br")
-            for section in about_sections:
-                if 'more information contact' in section:
-                    contact_title = section.split(",")[1]
-                    contact_name = section.split(",")[0]
-                    contact_name = contact_name.split("contact ")[1]
-                    contact_name = string_cleaner(contact_name)
-                    contact_title = string_cleaner(contact_title)
-                    print("Contact Name: " + contact_name +
-                          "\nContact Title: " + contact_title)
-                    break
-                else:
-                    contact_name = ""
-                    contact_title = ""
-        except:
-            contact_name = ""
-            contact_title = ""
-        try:
-            breadcrumbs = page_soup.find("ul", {"class": "bredcrump_list"})
-            crumbs = breadcrumbs.findAll("li")
-            category = crumbs[-2].text.strip()
-            category = category.replace(">>", "")
-            category = string_cleaner(category)
-            print("Category: " + category)
-        except:
-            category = ""
-    print("Validated: " + str(validated))
-    new_data_list.append(company_name)
-    new_data_list.append(phone_number)
-    new_data_list.append(reviews)
-    new_data_list.append(rating)
-    new_data_list.append(contact_name)
-    new_data_list.append(contact_title)
-    new_data_list.append(category)
-    new_data_list.append(validated)
-    return new_data_list
-
-
-def yellowpages_data_scraper(url, source_phone):
-    if url[-1].isdigit():
-        print(url)
-    else:
-        extra_string = url.split("/")[-1]
-        extra_string = "/" + extra_string
-        url = url.replace(extra_string, "")
-        print(url)
-    print(space)
-    validated = False
-    new_data_list = []
-    try:
-        driver.get("https://" + url)
-        html_page = driver.page_source
-        page_soup = soup(html_page, 'html.parser')
-        body = page_soup.find("body").text.strip()
-    except:
-        raw_input("Error on page loading")
-    try:
-        body = page_soup.find("body").text.strip()
-        if "Forbidden" in body:
-            raw_input("Forbidden, Please change IP an press Enter")
-    except:
-        pass
-    try:
-        company_name = page_soup.find("div", {"class": "sales-info"})
-        company_name = company_name.text.strip()
-        company_name = string_cleaner(company_name)
-        if 'add to favorites' in company_name:
-            company_name = company_name.split('add to favorites')[0]
-        print("Company Name: " + company_name)
-    except:
-        company_name = ""
-    try:
-        phone_number = page_soup.find("p", {"class": "phone"})
-        phone_number = phone_number.text.strip()
-        phone_number = remove_phone_format(phone_number)
-        if phone_number == source_phone:
-            validated = True
-        print("Phone Number: " + phone_number)
-    except:
-        phone = ""
-    try:
-        rating_containers = page_soup.findAll(
-            "meta", {"property": "og:description"})
-        for rating_container in rating_containers:
-            rating_container = str(rating_container)
-            if 'stars' in rating_container:
-                rating = rating_container.split(" stars on YP")[0]
-                rating = rating.replace('<meta content="Rated ', "")
-                rating = string_cleaner(rating)
-                print("Rating: " + rating)
-                break
-            else:
-                rating = ""
-    except:
-        rating = ""
-    try:
-        street_address = str(page_soup.find(
-            "script", {"type": "application/ld+json"}))
-        street_address = street_address.split('streetAddress":"')[1]
-        street_address = street_address.split('"')[0]
-        street_address = string_cleaner(street_address)
-        print("Street Address: " + street_address)
-    except:
-        street_address = ""
-    try:
-        city = str(page_soup.find("script", {"type": "application/ld+json"}))
-        city = city.split('addressLocality":"')[1]
-        city = city.split('"')[0]
-        city = string_cleaner(city)
-        print("City: " + city)
-    except:
-        city = ""
-    try:
-        state = str(page_soup.find("script", {"type": "application/ld+json"}))
-        state = state.split('addressRegion":"')[1]
-        state = state.split('"')[0]
-        state = string_cleaner(state)
-        print("State: " + state)
-    except:
-        state = ""
-    try:
-        zip_code = str(page_soup.find(
-            "script", {"type": "application/ld+json"}))
-        zip_code = zip_code.split('postalCode":"')[1]
-        zip_code = zip_code.split('"')[0]
-        zip_code = string_cleaner(zip_code)
-        print("Zip Code: " + zip_code)
-    except:
-        zip_code = ""
-    try:
-        email_address = page_soup.find(
-            "div", {"class": "business-card-footer"})
-        email_address = str(email_address)
-        email_address = email_address.split('mailto:')[1]
-        email_address = email_address.split('"')[0]
-        email_address = string_cleaner(email_address)
-        print("Email: " + email_address)
-    except:
-        email_address = ""
-    try:
-        years_container = page_soup.find("div", {"class": "number"})
-        years = years_container.text.strip()
-        years = string_cleaner(years)
-        print("Years in business " + years)
-    except:
-        years = ""
-    try:
-        reviews = str(page_soup.find(
-            "script", {"type": "application/ld+json"}))
-        reviews = reviews.split('reviewCount":')[1]
-        reviews = reviews.split('}')[0]
-        reviews = string_cleaner(reviews)
-        print("Reviews: " + reviews)
-    except:
-        reviews = ""
-    try:
-        website = str(page_soup.find(
-            "script", {"type": "application/ld+json"}))
-        website = website.split('"&nbsp;http://')[1]
-        website = website.split('&nbs')[0]
-        website = string_cleaner(website)
-        print("Website " + website)
-    except:
-        website = ""
-    try:
-        temp_phone_list = []
-        extra_phone_container = page_soup.find("dd", {"class": "extra-phones"})
-        extra_phone_container = extra_phone_container.findAll("span")
-    except:
-        alt_phone_1 = ""
-        alt_phone_2 = ""
-    else:
-        for p_number in extra_phone_container:
-            p_number = p_number.text.strip()
-            if ") " in p_number:
-                temp_phone_list.append(p_number)
-        alt_phone_1 = temp_phone_list[0]
-        alt_phone_1 = remove_phone_format(alt_phone_1)
-        if alt_phone_1 == source_phone:
-            validated = True
-        try:
-            alt_phone_2 = temp_phone_list[1]
-            alt_phone_2 = remove_phone_format(alt_phone_2)
-        except:
-            alt_phone_2 = ""
-        if alt_phone_2 == source_phone:
-            validated = True
-        print("Alt Phone Number 1: " + alt_phone_1 +
-              '\n' + "Alt Phone Number 2: " + alt_phone_2)
-    try:
-        category_container = page_soup.find("dd", {"class": "categories"})
-        category_container = category_container.text.strip()
-        category_container = category_container.split(',')
-    except:
-        category_1 = ""
-        category_2 = ""
-    else:
-        try:
-            category_1 = category_container[0]
-            category_1 = string_cleaner(category_1)
-            print("Category 1: " + category_1)
-        except:
-            category_1 = ""
-        try:
-            category_2 = category_container[1]
-            category_2 = string_cleaner(category_2)
-            print("Category 2: " + category_2)
-        except:
-            category_2 = ""
-    try:
-        accepted_payment = page_soup.find(
-            "dd", {"class": "payment"}).text.strip()
-        accepted_payment = string_cleaner(accepted_payment)
-        print("Accepted Payment: " + accepted_payment)
-    except:
-        accepted_payment = ""
-    try:
-        aka = page_soup.find("dd", {"class": "aka"}).text.strip()
-        aka = string_cleaner(aka)
-        print("Alias: " + aka)
-    except:
-        aka = ""
-    try:
-        business_info = page_soup.find("section", {"id": "business-info"})
-        business_info = str(business_info)
-        # print(business_info)
-        if 'mailto' in business_info:
-            alt_email = business_info.split('mailto:')[1]
-            alt_email = alt_email.split('"')[0]
-            alt_email = string_cleaner(alt_email)
-            print("Alt Email: " + alt_email)
-        else:
-            alt_email = ""
-    except:
-        alt_email = ""
-    print("Validated: " + str(validated))
-    try:
-        new_data_list.append(company_name)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(phone_number)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(rating)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(street_address)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(city)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(state)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(zip_code)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(email_address)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(years)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(reviews)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(website)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(alt_phone_1)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(alt_phone_2)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(category_1)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(category_2)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(accepted_payment)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(aka)
-    except:
-        new_data_list.append("error")
-    try:
-        new_data_list.append(alt_email)
-    except:
-        new_data_list.append("error")
-    new_data_list.append(validated)
-    return new_data_list
-
-
 def primary_sources_merge(dict_one, dict_two, dict_three, number_of_empty_lists):
     de_duped_lead_list = []
     unique_phone_list = []
@@ -1038,57 +383,57 @@ def check_for_no_results(dict_one, dict_two, dict_three):
     return empty_list_count
 
 
-def post_enhancement_data_scrape(enhanced_lead_data):
-    for enhanced_lead in enhanced_lead_data:
-        new_data_dictionaries = []
-        phone_number = enhanced_lead["phone"]
-        for value in enhanced_lead.values():
-            if 'yelp.com' in value:
-                already_scraped = False
-                for row in new_yelp_leads:
-                    yelp_url = row['yelp url']
-                    if value == yelp_url:
-                        already_scraped = True
-                        break
-                if already_scraped:
-                    pass
-                else:
-                    # temp_list = [value]
-                    yelp_data = yelp_data_scraper(driver, value, phone_number)
-                    new_yelp_data = yelp_data[0]
-                    new_data_dictionaries.append(new_yelp_data)
-            if 'bbb.org' in value:
-                already_scraped = False
-                for row in new_bbb_url_and_phones:
-                    bbb_url = row["bbb url"]
-                    if value == bbb_url:
-                        already_scraped = True
-                        break
-                if already_scraped:
-                    pass
-                else:
-                    new_bbb_data = bbb_data_scraper(value, phone_number)
-            if 'yellowpages.com' in value:
-                already_scraped = False
-                for row in new_yp_url_and_phones:
-                    try:
-                        yp_url = row[0]
-                    except:
-                        yp_url = row["yp url"]
-                    if value == yp_url:
-                        already_scraped = True
-                        break
-                if already_scraped:
-                    print("already_scraped")
-                else:
-                    new_yp_data = yellowpages_data_scraper(value, phone_number)
-            if 'manta.com' in value:
-                new_manta_data = manta_data_scraper(value, phone_number)
-            if 'mapquest.com' in value:
-                new_mapquest_data = mapquest_data_scraper(value, phone_number)
-            if 'chamberofcommerce.com' in value:
-                new_chamberofcommerce_data = chamberofcommerce_data_scraper(
-                    value, phone_number)
+# def post_enhancement_data_scrape(enhanced_lead_data):
+#     for enhanced_lead in enhanced_lead_data:
+#         new_data_dictionaries = []
+#         phone_number = enhanced_lead["phone"]
+#         for value in enhanced_lead.values():
+#             if 'yelp.com' in value:
+#                 already_scraped = False
+#                 for row in new_yelp_leads:
+#                     yelp_url = row['yelp url']
+#                     if value == yelp_url:
+#                         already_scraped = True
+#                         break
+#                 if already_scraped:
+#                     pass
+#                 else:
+#                     # temp_list = [value]
+#                     yelp_data = yelp_data_scraper(driver, value, phone_number)
+#                     new_yelp_data = yelp_data[0]
+#                     new_data_dictionaries.append(new_yelp_data)
+#             if 'bbb.org' in value:
+#                 already_scraped = False
+#                 for row in new_bbb_url_and_phones:
+#                     bbb_url = row["bbb url"]
+#                     if value == bbb_url:
+#                         already_scraped = True
+#                         break
+#                 if already_scraped:
+#                     pass
+#                 else:
+#                     new_bbb_data = bbb_data_scraper(driver, value, phone_number)
+#             if 'yellowpages.com' in value:
+#                 already_scraped = False
+#                 for row in new_yp_url_and_phones:
+#                     try:
+#                         yp_url = row[0]
+#                     except:
+#                         yp_url = row["yp url"]
+#                     if value == yp_url:
+#                         already_scraped = True
+#                         break
+#                 if already_scraped:
+#                     print("already_scraped")
+#                 else:
+#                     new_yp_data = yellowpages_data_scraper(driver, value, phone_number)
+#             if 'manta.com' in value:
+#                 new_manta_data = manta_data_scraper(value, phone_number)
+#             if 'mapquest.com' in value:
+#                 new_mapquest_data = mapquest_data_scraper(value, phone_number)
+#             if 'chamberofcommerce.com' in value:
+#                 new_chamberofcommerce_data = chamberofcommerce_data_scraper(
+#                     value, phone_number)
 
 
 def get_verticals_and_location_crawled():
